refactor : UserService, ArticleService, UserController 클래스에 @Autowired 어노테이션 제거
