+++ src/main/java/com/kakao/cafe/service/UserService.java
@@ -14,7 +14,6 @@ public class UserService {
 
     private UserRepository repository;
 
-    @Autowired
     public UserService(UserRepository repository) {
         this.repository = repository;
     }
